Fix patron specs for #post_file and #get_file
patron_spec.rb: specs fail when / is actually writable by the user
(which happens to be the case on Mac OS X). A file "/non_existing_file"
actually gets created in a spec for #get_file. A spec for #post_file
then expects that file to not be present and fails.