Make repeated keys in queries act like array queries all the time
Right now QueryMapper will normalize a query like:

`"abc=123&abc=456"` to `"abc=456"`

This means that if I interface with an API that expects
parameters passed that way, and I want to assert a_request to
have_been_made, I can't. The assertion will expect the form
`"abc=123&abc=456"`, but the captured request will be normalized to
`"abc=456"`.

This change simply makes the repetition of a key in the query turn that
particular parameter into an array, rather than writing over the
previous value.