Remove state from ManticoreAdapter
ManticoreAdapter was not thread-safe. Multiple threads would access
the same Client instance, which stored @uri, @method, and @options.
Multiple concurrent requests would cause data inconsistencies and
map responses to the incorrect request.

Fixing the thread-safety exposed several failing tests in the Manticore
specs. The gist of this is, Webmock allows you to stub with headers that
have Array values. e.g. `:headers => { 'Accept' => ['image/jpg', 'image/png'] }

However, you cannot pass an Array as a header value to the Manticore library.
It is the user's responsibility to concatenate this to a comma-separated string.

In order to match the configured stub, you have to naively split all
header values containing a comma into an Array, before attempting to match.

This feels wrong. However, I don't know a better way of doing it. My Webmock-fu
is weak.