Fix instance variable not initialized warnings
> /lib/util/version_checker.rb:83: warning: instance variable @max_major not initialized
> /lib/util/version_checker.rb:86: warning: instance variable @max_minor not initialized
> /lib/util/version_checker.rb:94: warning: instance variable @max_minor not initialized
> /lib/webmock/http_lib_adapters/excon_adapter.rb:41: warning: instance variable @original_excon_mock_default not initialized
> /lib/webmock/http_lib_adapters/excon_adapter.rb:43: warning: instance variable @stub not initialized

Since these are class methods there isn't an `initialize` for us to hook
into to set the variables. Instead we need to put ourselves in the class
object context to get the variables set. We do this early, before the
variables can be used, using `instance_exec`. We do not use the `eval`
version as we are not passing strings. The `_exec` version is available
in all supported Ruby versions.