Don't mutate the query, so immutable strings are acceptable
People who have opted in to the `frozen_string_literal: true` Ruby 2.3
option, which makes all string literals frozen/immutable by default,
will run into this an when attempting to use the `query` option in
`webmock`.

This option causes all string literals to be frozen. I've imitated this
in the test with an explicit `freeze` call on the argument. This will be
the default on Ruby 3. More details on this Ruby 2.3 option are at
https://wyeworks.com/blog/2015/12/1/immutable-strings-in-ruby-2-dot-3.

The `force_encoding` call here was causing a `can't modify frozen
String` error when this setting was enabled because, like the webmock
docs, I was passing in a string literal for the query.