Fix all the ruby 2.7 warnings coming from this code.
There's still a ton of broken tests... but this cleans the output down
to just straight test results. Most of the breakage that I've looked
at is down inside eventmachine. I don't know if that's the adapter's
fault or what. I'm not familiar with either project so I'll leave it
to the maintainers to do the actual debugging.

Signed-off-by: Ryan Davis <zenspider@chef.io>