Make global stubs thread-safe
This fixes an issue where testing multi-threaded code could sometimes result in incorrect responses being returned. Specifically in the case of VCR, since it seems to mock a nil response first before stubbing the real value, a nil response could sometimes be returned. This was because the hash being used by global stubs is not thread-safe.

I tried to write a test for this, but given the multi-threaded nature, and the fact it doesn't happen on MRI, it was difficult to find anything that failed. The existing tests do at least regression test it though.

closes https://github.com/bblimke/webmock/issues/907