Fix failing Async::HTTP specs due to Async API change
The `async` gem recently made a breaking API change that was released in
patch version 2.6.4.[^1] Following the change, async tasks that return
an  exception object will no longer raise that exception when `wait` is
called.

The `Async::HTTP` specs in webmock were written to leverage the old
behavior of `wait`. Since async 2.6.4 was released, these specs have
been failing in CI.

This commit fixes the failing specs by updating how `wait` is used, such
that exceptions are still raised as expected in async 2.6.4. This should
restore CI to a working state.

[^1]: https://github.com/socketry/async/pull/270