fix: Unexpected default rack.session
Some rack applications fail when the `rack.session` is already initialized,
but with an unexpected value. For example, in a Rails API-only application,
unexpected behavior is triggered when a `rack.session` is present.

This commit changes webmock, not to set a `rack.session` at all, similar
to how web servers do not set a session. The application or a middleware
sets them up if needed.

Fixes #985