Suspend JRuby CI matrix
This PR suspends JRuby CI matrix:

The issue has been reported at jruby/jruby#8606.
https://github.com/bblimke/webmock/actions/runs/13170199216/job/36759013080

The JRuby CI matrix has been failing for a long time, so pining to a version like
JRuby 9.4.10.0 will not avoid the error. Since the issue is likely complex and not easy to resolve,
it would be better to test with a version released after the resolution of jruby/jruby#8606.